pyproject.toml
[tool.poetry]
name = "chat-pdf"
version = "0.1.0"
description = ""
authors = ["PatLabCorp <patrick.alonso.fc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
ollama = "^0.3.3"
chromadb = "^0.5.11"
streamlit = "^1.39.0"
pdfplumber = "^0.11.4"
langchain = "^0.3.2"
langchain-core = "^0.3.9"
langchain-community = "^0.3.1"
langchain-text-splitters = "^0.3.0"
unstructured = "^0.15.13"
onnx = "^1.17.0"
instructor = "^1.5.1"
sentence-transformers = "2.2.2"
instructorembedding = "^1.0.1"
pi-heif = "^0.18.0"
unstructured-inference = "^0.7.37"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.8.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
